requirements.txt
bs4
discord.py
flask
gunicorn
hypixel.py[speed]
requests
sqlalchemy